# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LeadInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  /** @nullable */
  creditDebt?: string | null;
  /** @nullable */
  securedDebt?: string | null;
  /** @nullable */
  assets?: string | null;
  /** @nullable */
  income?: string | null;
}

export interface Lead {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  creditDebt?: string | null;
  /** @nullable */
  securedDebt?: string | null;
  /** @nullable */
  assets?: string | null;
  /** @nullable */
  income?: string | null;
  createdAt: string;
}

export interface AdminLoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface AdminLoginResult {
  ok: boolean;
}

export interface AdminSettings {
  /** @nullable */
  telegramBotToken: string | null;
  /** @nullable */
  telegramChatId: string | null;
  /** @nullable */
  telegramChatTitle: string | null;
}

export interface AdminSettingsUpdate {
  /** @nullable */
  telegramBotToken?: string | null;
}