# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: leads
    description: Consultation lead operations
  - name: admin
    description: Admin authentication and settings
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /leads:
    post:
      operationId: createLead
      tags: [leads]
      summary: Submit a consultation lead
      description: Public endpoint used by the lead form to submit a new consultation request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Invalid input
    get:
      operationId: listLeads
      tags: [leads]
      summary: List consultation leads
      description: Admin-only endpoint that returns all submitted leads, requires the X-Admin-Key header
      parameters:
        - name: X-Admin-Key
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
        "401":
          description: Unauthorized
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      description: Verifies admin username/password credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Unauthorized
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get admin settings
      description: Returns Telegram notification settings, requires the X-Admin-Key header
      parameters:
        - name: X-Admin-Key
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
        "401":
          description: Unauthorized
    put:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update admin settings
      description: Updates Telegram notification settings, requires the X-Admin-Key header
      parameters:
        - name: X-Admin-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSettingsUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
        "401":
          description: Unauthorized
  /admin/settings/detect-telegram-chat:
    post:
      operationId: detectTelegramChat
      tags: [admin]
      summary: Auto-detect Telegram group chat
      description: >-
        Uses the saved Telegram bot token to look up recent updates and automatically
        detect the group chat the bot has been added to. Saves the detected chat id
        and title to settings. Requires the X-Admin-Key header.
      parameters:
        - name: X-Admin-Key
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
        "400":
          description: No bot token saved, or no group chat could be detected
        "401":
          description: Unauthorized
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    LeadInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        creditDebt:
          type: ["string", "null"]
        securedDebt:
          type: ["string", "null"]
        assets:
          type: ["string", "null"]
        income:
          type: ["string", "null"]
      required:
        - name
        - phone
    Lead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        creditDebt:
          type: ["string", "null"]
        securedDebt:
          type: ["string", "null"]
        assets:
          type: ["string", "null"]
        income:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - createdAt
    AdminLoginInput:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
        - username
        - password
    AdminLoginResult:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    AdminSettings:
      type: object
      properties:
        telegramBotToken:
          type: ["string", "null"]
        telegramChatId:
          type: ["string", "null"]
        telegramChatTitle:
          type: ["string", "null"]
      required:
        - telegramBotToken
        - telegramChatId
        - telegramChatTitle
    AdminSettingsUpdate:
      type: object
      properties:
        telegramBotToken:
          type: ["string", "null"]