# lib/api-zod/src/generated/types/adminLoginInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}