# lib/api-zod/src/generated/types/adminSettings.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSettings {
  /** @nullable */
  telegramBotToken: string | null;
  /** @nullable */
  telegramChatId: string | null;
  /** @nullable */
  telegramChatTitle: string | null;
}